client-server lib working but using handler
The problem with handlers is that you will receive messages at another
thread, which may be a problem in rust because you can't access the
other thread's data without some sort of smart pointer. It'll be changed
to use channels instead allowing users to choose from which thread he/she will receive the
messages.